+++ internal/cli/test/condition.go
@@ -80,6 +80,12 @@ func (c *ConditionsMap) UnmarshalYAML(value *yaml.Node) error {
 				return fmt.Errorf("line %v: %v", v.Line, err)
 			}
 			cond = val
+		case "file_json_contains":
+			val := FileJSONContainsCondition("")
+			if err := v.Decode(&val); err != nil {
+				return fmt.Errorf("line %v: %v", v.Line, err)
+			}
+			cond = val
 		case "metadata_equals":
 			val := MetadataEqualsCondition{}
 			if err := v.Decode(&val); err != nil {
@@ -259,6 +265,30 @@ func (c FileJSONEqualsCondition) checkFrom(dir string, p *message.Part) error {
 	}
 
 	comparison := ContentJSONEqualsCondition(fileContent)
+	return comparison.Check(p)
+}
+
+//------------------------------------------------------------------------------
+
+// FileJSONContainsCondition is a string condition that tests the contents of the file
+// against the contents of a message using JSON comparison and is true if the expected
+// and actual documents are both valid JSON and the actual is a superset of the expected.
+type FileJSONContainsCondition string
+
+// Check this condition against a message part.
+func (c FileJSONContainsCondition) Check(p *message.Part) error {
+	return c.checkFrom("", p)
+}
+
+func (c FileJSONContainsCondition) checkFrom(dir string, p *message.Part) error {
+	relPath := filepath.Join(dir, string(c))
+
+	fileContent, err := os.ReadFile(relPath)
+	if err != nil {
+		return fmt.Errorf("failed to read comparison JSON file: %w", err)
+	}
+
+	comparison := ContentJSONContainsCondition(fileContent)
 	return comparison.Check(p)
 }