+++ internal/tracing/package.go
@@ -1,6 +1,3 @@
 // Package tracing implements utility functions for interacting with a global
-// tracing system. Currently this system uses the opentracing APIs. However,
-// eventually we will need to migrate to the new opentelemetry APIs, and
-// therefore in order to reduce disruption this package abstracts interaction
-// with these packages.
+// tracing system. Currently this system uses the opentelemetry APIs.
 package tracing